Code Reformat

+++ src/util_app.c
@@ -1,9 +1,10 @@
 #include "util_app.h"
 
+
 #define NRF_LOG_MODULE_NAME "[util]"
 
 int8_t app_disc_addr_check(uint8_t *p_data) {
-    for (int i = 0; i < APP.net.disc.cnt; i++) {
+    for (int8_t i = 0; i < APP.net.disc.cnt; i++) {
         if (!memcmp(APP.net.disc.peer[i].p_addr.addr, p_data, BLE_GAP_ADDR_LEN)) {
             LOG_D("ADDR FOUND!\r\n");
             return i;
